deploy/unicorn settings updated.

+++ lib/templates_install/config/deploy.rb
@@ -11,14 +11,14 @@ REPO_URL = ENV['GIT_REPOSITORY_URL']
 SECRET_KEY_BASE_KEY_NAME = 'SECRET_KEY_BASE'
 
 #################### Settings ####################
-set :application, ENV['PROJECT_NAME']
+set :application, ENV['APP_NAME']
 set :repo_url, REPO_URL
 set :repo_url_, ENV['GIT_REPOSITORY_URL']
 set :branch, 'release'
-set :deploy_to, "/var/www/projects/#{ENV['PROJECT_NAME']}"
+set :deploy_to, "/var/www/projects/#{ENV['APP_NAME']}"
 set :scm, :git
 set :format, :pretty
-set :log_level, :info
+set :log_level, :debug
 set :pty, true # default is false
 set :linked_files, fetch(:linked_files, []).push('config/database.yml', 'config/secrets.yml')
 
@@ -26,6 +26,7 @@ set :default_env, {
   rbenv_root:        "~/.rbenv",
   path:              "~/.rbenv/shims:~/.rbenv/bin:$PATH",
   project_name:      ENV['PROJECT_NAME'],
+  app_name:          ENV['APP_NAME'],
   database_name:     ENV['DATABASE_NAME'],
   database_host:     ENV['DATABASE_HOST'],
   database_port:     ENV['DATABASE_PORT'],
@@ -39,7 +40,7 @@ set :rbenv_path, '~/.rbenv'
 
 set :keep_releases, 1 # default is 5
 set :unicorn_config_path, "config/unicorn.rb"
-set :unicorn_pid, "/var/run/#{ENV['PROJECT_NAME']}/unicorn/unicorn.pid"
+set :unicorn_pid, "/var/run/#{ENV['APP_NAME']}/unicorn/unicorn.pid"
 
 
 #################### Tasks ####################
@@ -61,7 +62,9 @@ namespace :db do
     on roles :db do
       with rails_env: fetch(:rails_env) do
         within current_path do
-          execute :bundle, :exec, :rake, 'db:create'
+          if test "[ ! psql -h #{fetch(:database_host)} -U #{fetch(:database_user)} -c \"select * from pg_database where datname='#{fetch(:database_name)}'\" | grep -q #{fetch(:database_name)} ]"
+            execute :bundle, :exec, :rake, 'db:create'
+          end
         end
       end
     end
@@ -72,7 +75,9 @@ namespace :db do
     on roles :db do
       with rails_env: fetch(:rails_env) do
         within current_path do
-          execute :bundle, :exec, :rake, 'db:drop'
+          if test "[ ! psql -h #{fetch(:database_host)} -U #{fetch(:database_user)} -c \"select * from pg_database where datname='#{fetch(:database_name)}'\" | grep -q #{fetch(:database_name)} ]"
+            execute :bundle, :exec, :rake, 'db:drop'
+          end
         end
       end
     end
@@ -116,7 +121,6 @@ namespace :deploy do
       end
     end
   end
-
   
   task :upload do
     on roles(:app) do |host|
@@ -131,11 +135,13 @@ namespace :deploy do
   desc 'Set secret_key_base for rails'
   task :set_key do
     on roles(:app) do
-      if test " [ ! -d #{shared_path}/env] "
-        execute "mkdir -p #{shared_path}/env"
+      within current_path do
+        if test "[ ! -d #{shared_path}/env ]"
+          execute "mkdir -p #{shared_path}/env"
+        end
+        execute "echo #{SECRET_KEY_BASE_KEY_NAME}: #{ENV['SECRET_KEY_BASE']} > #{shared_path}/env/env.yml"
+        execute "echo #{SECRET_KEY_BASE_KEY_NAME + "_STG"}: #{ENV['SECRET_KEY_BASE_STG']} >> #{shared_path}/env/env.yml"
       end
-      execute "echo #{SECRET_KEY_BASE_KEY_NAME}: #{ENV[SECRET_KEY_BASE_KEY_NAME]} > #{shared_path}/env/env.yml"
-      execute "echo #{SECRET_KEY_BASE_KEY_NAME + "_STG"}: #{ENV[SECRET_KEY_BASE_KEY_NAME + "_STG"]} >> #{shared_path}/env/env.yml"
     end
   end
   
@@ -156,10 +162,11 @@ namespace :deploy do
   
   before 'deploy:upload', 'deploy:directories'
   before 'deploy:starting', 'deploy:upload'
-  after 'deploy:finishing', 'deploy:set_key'
   after 'deploy:finishing', 'deploy:cleanup'
-  after 'deploy:publishing', 'deploy:restart'
-  after 'deploy:publishing', 'deploy:update_schema'
+  after 'deploy:publishing', 'deploy:set_key'
+  after 'deploy:set_key', 'db:create'
+  after 'db:create', 'deploy:update_schema'
+  after 'deploy:update_schema', 'deploy:restart'
 
 end
